backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.41
starlette==0.46.2
cryptography==45.0.2
bcrypt==4.3.0
python-multipart==0.0.20
pydantic==2.11.4
alembic==1.16.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
httpx==0.28.1
redis==6.1.0
celery==5.5.2
flower==2.0.1
python-magic==0.4.27
pillow==11.2.1

